:art: [ Refactory,Fix ] CardListPage custom hook으로 핸들러 함수 등 정리 및 useRequestApi 버그 수정
refactory, fix

(cherry picked from commit a8d5006bd4fe74d41380fff822c2450278e9836c)
